rotating objects
if you assign selected objects to the player, then press F to inspect, you can rotate them with qweasd. un inspecting will reset to 0,0,0 rotation.

Make raycast to detect when looking at a selectable object and assign it to player on click.

make moving the object with the player a thing.

Make UI telling you how to rotate and enter inspection.

make checking area where you place the option for the puzzle and it checks if it is correct.